+++ src/02-tagger/tiny.js
@@ -4,6 +4,10 @@ const lexiconStep = require('./01-init/01-lexicon')
 const smallTagger = function (doc) {
   let terms = doc.termList()
   lexiconStep(terms, doc.world)
+  // run any user-given tagger functions
+  doc.world.taggers.forEach(fn => {
+    fn(doc)
+  })
   return doc
 }
 module.exports = smallTagger